Logs de fluxo das requisições Http

+++ 04_RestWithASPNET10_ScaffoldViaVisualStudio/RestWithASPNET10/RestWithASPNET10/Services/IPersonService.cs
@@ -5,7 +5,7 @@ namespace RestWithASPNET10.Services;
 public interface IPersonService
 {
     Person Create(Person person);
-    Person FindById(int id);
+    Person FindById(long id);
     List<Person> FindAll();
     Person Update(Person person);
     void Delete(long id);
