+++ 04_RestWithASPNET10_ScaffoldViaVisualStudio/RestWithASPNET10/RestWithASPNET10/Services/Impl/PersonServicesImpl.cs
@@ -15,7 +15,7 @@ public class PersonServicesImpl : IPersonService
         return _context.Persons.ToList();
     }
 
-    public Person FindById(int id)
+    public Person FindById(long id)
     {
         return _context.Persons.Find(id);
     }
